transfer private datas to the end of the Maybe class

+++ task5_Maybe/src/test/java/ru/spbau/karlina/task5_maybe/MaybeTest.java
@@ -70,7 +70,7 @@ public class MaybeTest {
     @Test
     public void mapStringLengthTest() throws Exception {
         Maybe<String> maybe = Maybe.just("string");
-        assertEquals(new Integer(5), maybe.map(String::length).get());
+        assertEquals(new Integer(6), maybe.map(String::length).get());
     }
 
     @Test
